Refactor : ErrorMessage 클래스의 위치를 utils에서 exception 패키지로 이동
