+++ src/main/java/com/codessquad/qna/exception/EntityNotFoundException.java
@@ -1,7 +1,5 @@
 package com.codessquad.qna.exception;
 
-import com.codessquad.qna.utils.ErrorMessage;
-
 public class EntityNotFoundException extends RuntimeException {
     public EntityNotFoundException(ErrorMessage errorMessage) {
         super(errorMessage.getErrorMessage());
