+++ src/main/java/com/codessquad/qna/exception/IllegalUserAccessException.java
@@ -1,7 +1,5 @@
 package com.codessquad.qna.exception;
 
-import com.codessquad.qna.utils.ErrorMessage;
-
 public class IllegalUserAccessException extends RuntimeException {
     public IllegalUserAccessException() {
         super(ErrorMessage.INVALID_USER.getErrorMessage());
